requirements.txt
Django==5.2.12
google-generativeai==0.8.6
requests==2.32.3
python-dotenv==1.0.1
markdown==3.10.2
